리팩토링: Money 값 객체 도입 + 응답 객체·인프라 정리 (#69)
* mini-coupang: Qdrant 컴포즈 서비스와 Python ML 모듈 제거

하이브리드 검색 인프라(Qdrant + bge-m3 gRPC) 정리. shared/docker/docker-compose.yml에서 qdrant 서비스와 qdrant-data 볼륨 제거, mini-coupang/ml/ 디렉터리 전체 삭제.

* mini-coupang: docker compose 파일을 시나리오별 3종으로 재구성

기존 4개 파일(base + observability + integration-test + override)을 용도별로 재배치.
각 서비스 모듈 root에 단독 개발용 dev compose, integration-test root에 통합 테스트 풀 스택,
shared/docker에 스테이징 유사 local 풀 스택. Kafka 사용 서비스는 Kafka UI, Redis 사용 서비스는
RedisInsight를 함께 포함하도록 의존성 매트릭스 반영.

* mini-coupang: docker-compose-local.yml 을 KRaft 3-broker 클러스터로 확장

local 풀 스택 한정으로 Kafka broker 3대 (kafka-1/2/3) 구성. RF=3, MIN_ISR=2 로
leader fail-over 와 ISR shrink·expand 시나리오 학습이 가능한 환경. 호스트 접근 포트
29092/29192/29292 로 분리, 모든 서비스와 kafka-ui 의 bootstrap-servers 를 멀티 브로커
리스트로 갱신.

5개 서비스 application.yaml 의 producer 설정에 acks=all, enable.idempotence=true,
max.in.flight.requests.per.connection=5, retries=10 을 추가해 다중 브로커 환경에서
정합성 보장이 실제로 동작하도록 변경. dev / integration-tests 는 단일 브로커 유지.

* docs: integration-test 실행 명령을 :it 태스크로 정정

build.gradle 에서 test 태스크는 enabled=false 이고 실제 통합 테스트는
별도의 it 태스크에서 실행됨. README 의 잘못된 :test 명령을 :it 으로 변경.

* mini-coupang: CommonResponse를 성공·실패 통합 응답 객체로 재작성

ErrorResponse를 흡수해 (code, message, data) 단일 record로 정리.
GlobalExceptionHandler 3종이 CommonResponse.failure(...)를 반환하도록 갱신.

* mini-coupang: Money 값 객체 + 단위 테스트 추가

@Embeddable record. 생성 시 BigDecimal HALF_UP 으로 10원 단위 반올림.
add/subtract/multiply(long), 비교 연산, IllegalArgumentException 음수 검증.
common-modules 에 AssertJ 추가 후 29 케이스 단위 테스트 통과.

* mini-coupang: Money 저장 타입을 BigDecimal 로 정리

ZERO 상수만 노출 (zero() 메서드는 lint 권고대로 제거).
반올림 전략은 compact constructor 인라인 + 주석으로 단순화.
중복인 won() 팩토리 제거.

* mini-coupang: Payment 도메인을 Money 값 객체로 마이그레이션

Payment 엔티티의 amount 필드를 long 에서 Money 로 교체.
@Embedded + @AttributeOverride 로 기존 amount 컬럼 유지.
PaymentConfirmService 가 command.amount() (long) → Money.of(...) 단일 변환.
경계 (controller dto, command, port dto, Kafka payload) 는 long 유지.

* mini-coupang: Order 도메인을 Money 값 객체로 마이그레이션

Order.totalAmount, OrderItem.unitPrice, OrderItem.lineAmount 를 Money 로 교체.
@Embedded + @AttributeOverride 로 기존 컬럼명 유지.
addItem/forOrder 시그니처는 Money unitPrice 를 받고, OrderService 가
snapshot.unitPrice() (long) 을 Money.of(...) 로 단일 변환.
lineAmount 계산은 unitPrice.multiply(quantity), totalAmount 누적은 .add(...).
Money.toLong() 헬퍼 추가, Result/Snapshot 3종이 .toLong() 으로 wire 호환 유지.
경계 (controller dto, command, port dto, Kafka payload) 는 long 그대로.

* mini-coupang: Product 도메인을 Money 값 객체로 마이그레이션 + Money 확장

Money 에 multiply(BigDecimal), subtractOrZero, min/max, sum(Iterable) 추가.
정률 할인, 가격 클램프, 캡, 합계 등 향후 쿠폰/할인 로직 빌딩 블록으로 사용.

Product.basePrice, ProductOption.additionalPrice 를 Money 로 교체 (@Embedded).
Product.create / addOption / addDefaultOption 시그니처에서 Money 를 받는다.
SellerProductService 가 command Long → Money.of(...) 단일 변환,
OptionDetailService 의 unitPrice 계산은 base.add(additional).toLong().
ProductRepository 의 JPQL 가격 필터 경로를 p.basePrice.amount 로 갱신.
경계 (controller dto, command, result) 는 long 그대로.

+++ mini-coupang/order-server/src/main/java/com/deepblog/order/application/result/ConfirmOrderResult.java
@@ -18,7 +18,7 @@ public record ConfirmOrderResult(
             order.getId(),
             order.getMemberId(),
             order.getStatus().name(),
-            order.getTotalAmount(),
+            order.getTotalAmount().toLong(),
             paymentId
         );
     }
