+++ mini-coupang/order-server/src/main/java/com/deepblog/order/application/result/PendingOrderSnapshot.java
@@ -23,7 +23,7 @@ public record PendingOrderSnapshot(
             order.getMemberId(),
             item.getOptionId(),
             item.getQuantity(),
-            order.getTotalAmount()
+            order.getTotalAmount().toLong()
         );
     }
 }
