+++ mini-coupang/order-server/src/main/java/com/deepblog/order/application/result/PrepareOrderResult.java
@@ -20,7 +20,7 @@ public record PrepareOrderResult(
             order.getId(),
             order.getMemberId(),
             order.getStatus().name(),
-            order.getTotalAmount(),
+            order.getTotalAmount().toLong(),
             optionId,
             quantity
         );
