+++ mini-coupang/product-server/src/main/java/com/deepblog/product/application/result/RegisterProductResult.java
@@ -19,7 +19,7 @@ public record RegisterProductResult(
             product.getSellerId(),
             product.getCategoryId(),
             product.getName(),
-            product.getBasePrice(),
+            product.getBasePrice().toLong(),
             product.getStatus().name(),
             product.getOptions().size(),
             product.getImages().size()
